Extracted InstanceData to its own file
